+++ quantities/Globals.cs
@@ -1,5 +1,3 @@
-global using Quantities.Numerics;
-
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("quantities.test")]
