+++ quantities/measures/IInjector.cs
@@ -1,3 +1,4 @@
+using Quantities.Numerics;
 using Quantities.Prefixes;
 using Quantities.Units;
 using static Quantities.Extensions;
