+++ quantities/prefixes/Build.cs
@@ -1,3 +1,4 @@
+using Quantities.Numerics;
 using static Quantities.Extensions;
 
 namespace Quantities.Prefixes;
